feat: Add game dropdown (#1450)
* wip: changed AddGameForm and Game model

* fix tests

* saving SQL query just in case, though it's not
fully working yet

* fixed SQL to delete avatars as well

* format migration

* move AddGameForm to portal

* updated ci

* Merge branch 'development' into add-game-dropdown

* fixed worksheet_id in cypress

* renamed migration

* bring avatar creator back
