+++ .github/workflows/cicd.yml
@@ -27,6 +27,7 @@ jobs:
         run: |
           pip install pipenv
           pipenv install --dev --system
+          pip install -U git+https://github.com/ocadotechnology/codeforlife-portal.git@add-game-dropdown  # Remove after merge
           cd game_frontend
           yarn --frozen-lockfile
       - name: Run Javascript tests
