+++ aimmo/templatetags/players_utils.py
@@ -14,11 +14,9 @@ def game_dropdown_list(context, base_url):
 def get_user_playable_games(context, base_url):
     user = context.request.user
     if logged_in_as_student(user):
-        playable_games = user.userprofile.student.class_field.games_for_class.all()
+        playable_games = user.userprofile.student.class_field.game
     elif logged_in_as_teacher(user):
-        playable_games = Game.objects.none()
-        for klass in user.userprofile.teacher.class_teacher.all():
-            playable_games |= klass.games_for_class.all()
+        playable_games = Game.objects.filter(game_class__teacher=user.userprofile.teacher)
     else:
         playable_games = Game.objects.none()
     return {
